requirements.txt
aiogram>=3.0
pytesseract
opencv-python
pandas
openpyxl
Pillow
numpy
scipy
torch
torchvision
ultralytics
easyocr
transformers